Refactored plotting functions to their own file
